requirements.txt
Django==1.11.4
djangorestframework==3.6.4
httpie==0.9.9
django-redis==4.8.0
model-mommy==1.4.0
django-extensions==1.9.0
ipython==6.1.0